requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
openai==1.10.0
redis==5.0.1
psycopg2-binary==2.9.9
numpy==1.26.3
scikit-learn==1.4.0
httpx==0.26.0